// shared/types/User.ts
import { ObjectId } from "mongodb";

export type User = {
	/**
	 * user id
	 *
	 * @name _id
	 */
	_id?: ObjectId;
	/**
	 * role of the user
	 *
	 * @name _role
	 */
	_role: "user" | "moderator" | "administrator";
	/**
	 * metadata
	 *
	 * @name _metadata
	 */
	_metadata: {
		views: number; // account views
	};
	/**
	 * account username
	 *
	 * @name username
	 */
	username: string;
	/**
	 * salted SHA256 password hash
	 *
	 * @name password
	 */
	password: string;
	/**
	 * account email address
	 *
	 * @name emailAddress
	 */
	emailAddress?: string;
	/**
	 * account display name
	 *
	 * @name displayName
	 */
	displayName?: string;
	/**
	 * account avatar url
	 *
	 * @name avatarUrl
	 */
	avatarUrl?: string;
	/**
	 * account banner url
	 *
	 * @name bannerUrl
	 */
	bannerUrl?: string;
	/**
	 * list of follower user ids
	 *
	 * @name followers
	 */
	followers?: string[];
	/**
	 * list of following user ids
	 *
	 * @name following
	 */
	following?: string[];
	/**
	 * number of posts, generated by server
	 *
	 * @name postCount
	 */
	postCount?: number;
	/**
	 * email verification state
	 *
	 * @name isEmailVerified
	 */
	isEmailVerified?: boolean;
	/**
	 * account creation timestamp
	 *
	 * @name timestamp
	 */
	timestamp: number;
	/**
	 * account lockout end timestamp, for moderation purposes
	 *
	 * @name timestamp
	 */
	lockout?: number;
};